Start enabling compiler specific extra warnings (#219)
* Start enabling compiler specific extra warnings

* Clean up Clang warnings

* Gcc warning fixes

* Silence application warning

* Fix warnings

* Remove explicit commandline parsing

* Futher warning fixes

* Address tempfile warnings

* Add basic application commandline coonfig test for coverage

* Macro correction

* Unused parameter removal

* Fix signed to unsigned comparisons

* Remove unused param and disabple unreachable warning in tests

* Remove unused param

* Guard compiler specific pragmas

* Fix warnings

* Futher warning fixes

* Proprocessor correction

* Enable extra warnigns

* Clang only warning

* Remove extra semi-colons

* Fix warnings

* Fix warning in indirect

* Update indirect tests

* Add comments only

* Loop brakets to correctly swallow the semicolon

* Remove warnings

* Release assert is void statement

* Initialise tracked base

* Ensure expected number of bytes read

* Ensure tracked is used in copy constructor

* Enable more warning and clear up warnings

* Remove shadowed variable warnings

* Document warning flag

* Fix warning in external libraries

+++ libraries/application/tests/po/config.tests.cpp
@@ -17,7 +17,7 @@ TEST_CASE("Ensure options parsing of standard application config", "[morpheus.ap
         WHEN("Parsing valid parameters")
         {
             std::array const cliOptions = { "dummyProgram.exe", "--logging-enabled", "true", "--log-append", "false", "--log-level", "debug", "--logfile-path", "." };
-            auto const result = parseProgramOptions(cliOptions.size(), cliOptions.data(), HelpDocumentation{}, confg);
+            auto const result = parseProgramOptions(static_cast<int>(cliOptions.size()), cliOptions.data(), HelpDocumentation{}, confg);
 
             THEN("Expect no error results and valid values extracted")
             {
